// types/stop-reason.ts
export type StopReason =
  | "COMPLETED"
  | "USER_STOP"
  | "AUTOPILOT_DISABLED"
  | "WAITING_FOR_APPROVAL"
  | "WAITING_FOR_DIRECTORY_TRUST"
  | "WAITING_FOR_USER_INPUT"
  | "WAITING_FOR_EXTERNAL_RESOURCE"
  | "PERMISSION_DENIED"
  | "POLICY_BLOCKED"
  | "PATH_OR_TOOL_NOT_ALLOWED"
  | "CONTEXT_EXHAUSTED"
  | "RETRY_EXHAUSTED"
  | "NON_PROGRESS_LIMIT"
  | "UNRECOVERABLE_ERROR"
  | "AMBIGUOUS_STATE_REQUIRES_ESCALATION";